move pointers to generic folder
