--- syft/frameworks/torch/pointers/__init__.py
@@ -1,7 +1,0 @@
-from syft.frameworks.torch.pointers.object_pointer import ObjectPointer
-from syft.frameworks.torch.pointers.callable_pointer import create_callable_pointer
-from syft.frameworks.torch.pointers.callable_pointer import CallablePointer
-from syft.frameworks.torch.pointers.pointer_tensor import PointerTensor
-from syft.frameworks.torch.pointers.object_wrapper import ObjectWrapper
-
-__all__ = ["ObjectPointer", "CallablePointer", "create_callable_pointer", "PointerTensor"]
